+++ common/src/main/java/dev/xylonity/bonsai/clockwork/registry/ClockworkItems.java
@@ -25,8 +25,8 @@ public class ClockworkItems {
     public static final ResourceEntry<Item> CROSSBOW_BARREL = ITEMS.register("crossbow_barrel", () -> new Item(new Item.Properties()));
     public static final ResourceEntry<Item> CLOCKWORK_ARROW = ITEMS.register("clockwork_arrow", () -> new ClockworkArrow(new Item.Properties()));
 
-    public static final ResourceEntry<Item> BARREL_CROSSBOW = ITEMS.register("barrel_crossbow", () -> new BarrelCrossbow(new Item.Properties().stacksTo(1)));
-    public static final ResourceEntry<Item> SCOPE_CROSSBOW = ITEMS.register("scope_crossbow", () -> new ScopeCrossbow(new Item.Properties().stacksTo(1)));
+    public static final ResourceEntry<Item> BARREL_CROSSBOW = ITEMS.register("barrel_crossbow", () -> new BarrelCrossbow(new Item.Properties().stacksTo(1).durability(ClockworkConfig.BARREL_CROSSBOW_DEFAULT_DURABILITY)));
+    public static final ResourceEntry<Item> SCOPE_CROSSBOW = ITEMS.register("scope_crossbow", () -> new ScopeCrossbow(new Item.Properties().stacksTo(1).durability(ClockworkConfig.SCOPE_CROSSBOW_DEFAULT_DURABILITY)));
 
     public static final ResourceEntry<Item> CLOCKWORK_POTION_SPRAYER = ITEMS.register("clockwork_potion_sprayer", () -> new PotionSprayer(new Item.Properties().stacksTo(1)));
     public static final ResourceEntry<Item> CLOCKWORK_FLAMETHROWER = ITEMS.register("clockwork_flamethrower", () -> new FlamethrowerItem(new Item.Properties().stacksTo(1).durability(ClockworkConfig.FLAMETHROWER_DEFAULT_DURABILITY)));